software/requirements.txt
opencv-python
PyQt5
pillow
matplotlib
pyjson